melositos-chatbot/chatbot-service/requirements.txt
fastapi
uvicorn
pypdf
chromadb
sentence-transformers